검색 통계 개발중 - 스코어 계산

+++ src/main/java/com/portfolio/aips/project/elastic_search/archive/service/archive_el_trending_search_log/command/RedisSaveCommand.java
@@ -6,5 +6,5 @@ import com.portfolio.aips.project.elastic_search.archive.service.archive_el_tren
 import com.portfolio.aips.project.elastic_search.archive.service.archive_el_trending_search_log.repo.ArchiveELTrendingSearchLogRedisRepository;
 
 // param: createDateTime 새로운 키워드 날짜
-public record RedisSaveCommand(String keyword, String docCount, String createDateTime, SearchDateRange range, int score)  {
+public record RedisSaveCommand(String keyword, String docCount, String createDateTime, SearchDateRange range, double score)  {
 }
